feat: refactor nx cli commands, throw out deprecated or outdated stuff
