+++ apps/vscode/package.json
@@ -75,16 +75,6 @@
           "group": "7_modification@-1"
         },
         {
-          "when": "isNxWorkspace && explorerResourceIsFolder && resourcePath in nxAppsDir && config.nxConsole.enableGenerateFromContextMenu && nx.hasApplicationGenerators",
-          "command": "nx.generate.ui.app.fileexplorer",
-          "group": "2_workspace"
-        },
-        {
-          "when": "isNxWorkspace && explorerResourceIsFolder && resourcePath in nxLibsDir && config.nxConsole.enableGenerateFromContextMenu && nx.hasLibraryGenerators",
-          "command": "nx.generate.ui.lib.fileexplorer",
-          "group": "2_workspace"
-        },
-        {
           "when": "isNxWorkspace",
           "group": "2_workspace@5",
           "command": "nx.graph.focus"
@@ -175,14 +165,6 @@
           "when": "false"
         },
         {
-          "command": "nx.generate.ui.app.fileexplorer",
-          "when": "false"
-        },
-        {
-          "command": "nx.generate.ui.lib.fileexplorer",
-          "when": "false"
-        },
-        {
           "command": "nxConsole.runTask",
           "when": "false"
         },
@@ -193,14 +175,6 @@
         {
           "command": "nxConsole.revealInExplorer",
           "when": "false"
-        },
-        {
-          "command": "nx.generate.ui.app",
-          "when": "isNxWorkspace && nx.hasApplicationGenerators"
-        },
-        {
-          "command": "nx.generate.ui.lib",
-          "when": "isNxWorkspace && nx.hasLibraryGenerators"
         },
         {
           "command": "nxConsole.migrateAngularCliToNx",
@@ -268,14 +242,6 @@
         },
         {
           "command": "nx.affected.lint",
-          "when": "isNxWorkspace"
-        },
-        {
-          "command": "nx.affected.apps",
-          "when": "isNxWorkspace"
-        },
-        {
-          "command": "nx.affected.libs",
           "when": "isNxWorkspace"
         },
         {
@@ -450,16 +416,6 @@
       },
       {
         "category": "Nx",
-        "title": "affected apps",
-        "command": "nx.affected.apps"
-      },
-      {
-        "category": "Nx",
-        "title": "affected libs",
-        "command": "nx.affected.libs"
-      },
-      {
-        "category": "Nx",
         "title": "list",
         "command": "nx.list"
       },
@@ -492,26 +448,6 @@
         "category": "Nx",
         "title": "Run Target from selected file",
         "command": "nx.run.fileexplorer"
-      },
-      {
-        "category": "Nx",
-        "title": "generate application",
-        "command": "nx.generate.ui.app"
-      },
-      {
-        "category": "Nx",
-        "title": "generate library",
-        "command": "nx.generate.ui.lib"
-      },
-      {
-        "category": "Nx",
-        "title": "Nx generate application",
-        "command": "nx.generate.ui.app.fileexplorer"
-      },
-      {
-        "category": "Nx",
-        "title": "Nx generate library",
-        "command": "nx.generate.ui.lib.fileexplorer"
       },
       {
         "category": "Nx",
