+++ libs/vscode/nx-cli-quickpicks/README.md
@@ -1,0 +1,7 @@
+# vscode-nx-cli-quickpicks
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test vscode-nx-cli-quickpicks` to execute the unit tests via [Jest](https://jestjs.io).
