+++ libs/vscode/nx-cli-quickpicks/src/lib/select-re-move-generator.ts
