+++ libs/vscode/tasks/src/index.ts
@@ -1,6 +1,3 @@
 export * from './lib/cli-task-provider';
 export * from './lib/nx-task-commands';
 export * from './lib/cli-task-commands';
-export * from './lib/cli-task-quick-pick-item';
-export * from './lib/select-generator';
-export * from './lib/select-flags';
