--- libs/vscode/tasks/src/lib/cli-task-flag-quick-pick-item.ts
@@ -1,11 +1,0 @@
-import { Option } from '@nx-console/shared/schema';
-import { QuickPickItem } from 'vscode';
-
-export class CliTaskFlagQuickPickItem implements QuickPickItem {
-  constructor(
-    readonly flagName: string,
-    readonly detail: string = '',
-    readonly option: Option,
-    readonly label: string
-  ) {}
-}
