+++ libs/vscode/tasks/src/lib/nx-task-commands.ts
@@ -1,3 +1,4 @@
+import { join } from 'path';
 import {
   commands,
   ExtensionContext,
@@ -6,18 +7,19 @@ import {
   tasks,
   window,
 } from 'vscode';
-import { join } from 'path';
 
-import { Option, OptionType } from '@nx-console/shared/schema';
+import { readAndParseJson } from '@nx-console/shared/file-system';
+import { getNxWorkspace } from '@nx-console/vscode/nx-workspace';
 import {
   getTelemetry,
   resolveDependencyVersioning,
 } from '@nx-console/vscode/utils';
-import { getNxWorkspace } from '@nx-console/vscode/nx-workspace';
-import { readAndParseJson } from '@nx-console/shared/file-system';
 
+import {
+  selectAffectedFlags,
+  selectRunManyFlags,
+} from '@nx-console/vscode/nx-cli-quickpicks';
 import { NxTask } from './nx-task';
-import { selectFlags } from './select-flags';
 
 export function registerNxCommands(context: ExtensionContext) {
   context.subscriptions.push(
@@ -29,10 +31,10 @@ export function registerNxCommands(context: ExtensionContext) {
       }
       promptForAffectedFlags(target);
     }),
-    ...['apps', 'build', 'e2e', 'libs', 'lint', 'test'].map((command) =>
-      commands.registerCommand(`nx.affected.${command}`, () => {
-        getTelemetry().featureUsed(`nx.affected.${command}`);
-        promptForAffectedFlags(command);
+    ...['build', 'e2e', 'lint', 'test'].map((target) =>
+      commands.registerCommand(`nx.affected.${target}`, () => {
+        getTelemetry().featureUsed(`nx.affected.${target}`);
+        promptForAffectedFlags(target);
       })
     )
   );
@@ -84,145 +86,7 @@ async function promptForTarget(): Promise<string | undefined> {
   return window.showQuickPick(validTargets);
 }
 
-const AFFECTED_OPTIONS: Option[] = [
-  {
-    name: 'base',
-    type: OptionType.String,
-    isRequired: false,
-    description: 'Base of the current branch (usually main)',
-  },
-  {
-    name: 'head',
-    type: OptionType.String,
-    isRequired: false,
-    description: 'Latest commit of the current branch (usually HEAD)',
-  },
-  {
-    name: 'parallel',
-    type: OptionType.Boolean,
-    description: 'Parallelize the command',
-    isRequired: false,
-    default: 'false',
-  },
-  {
-    name: 'maxParallel',
-    type: OptionType.Number,
-    description: 'Max number of parallel processes',
-    isRequired: false,
-    default: 3,
-  },
-  {
-    name: 'only-failed',
-    type: OptionType.Boolean,
-    description: 'Isolate projects which previously failed',
-    isRequired: false,
-    default: 'false',
-  },
-  {
-    name: 'all',
-    type: OptionType.Boolean,
-    description: 'All projects',
-    isRequired: false,
-  },
-  {
-    name: 'configuration',
-    type: OptionType.String,
-    isRequired: false,
-    description:
-      'This is the configuration to use when performing tasks on projects',
-  },
-  {
-    name: 'exclude',
-    type: OptionType.String,
-    isRequired: false,
-    description: 'Exclude certain projects from being processed',
-  },
-  {
-    name: 'files',
-    type: OptionType.Array,
-    isRequired: false,
-    description: 'Manually specify changed files, delimited by commas',
-  },
-  {
-    name: 'skip-nx-cache',
-    type: OptionType.Boolean,
-    description:
-      'Rerun the tasks even when the results are available in the cache',
-    isRequired: false,
-    default: false,
-  },
-  {
-    name: 'verbose',
-    type: OptionType.Boolean,
-    isRequired: false,
-    description: 'Print additional error stack trace on failure',
-    default: false,
-  },
-].map((v) => ({ ...v, aliases: [] }));
-
-const RUN_MANY_OPTIONS: Option[] = [
-  {
-    name: 'all',
-    type: OptionType.Boolean,
-    description: 'All projects',
-    isRequired: false,
-  },
-  {
-    name: 'parallel',
-    type: OptionType.Number,
-    description: 'Max number of parallel processes [default is 3]',
-    isRequired: false,
-    default: 3,
-  },
-  {
-    name: 'only-failed',
-    type: OptionType.Boolean,
-    description: 'Isolate projects which previously failed',
-    isRequired: false,
-    default: 'false',
-  },
-  {
-    name: 'configuration',
-    type: OptionType.String,
-    description:
-      'This is the configuration to use when performing tasks on projects',
-    isRequired: false,
-  },
-  {
-    name: 'skip-nx-cache',
-    type: OptionType.Boolean,
-    description:
-      'Rerun the tasks even when the results are available in the cache',
-    default: false,
-    isRequired: false,
-  },
-  {
-    name: 'with-deps',
-    type: OptionType.Boolean,
-    description:
-      'Include dependencies of specified projects when computing what to run',
-    isRequired: false,
-    default: false,
-  },
-  {
-    name: 'exclude',
-    type: OptionType.String,
-    isRequired: false,
-    description: 'Exclude certain projects from being processed',
-  },
-  {
-    name: 'verbose',
-    type: OptionType.Boolean,
-    description: 'Print additional error stack trace on failure',
-    isRequired: false,
-    default: false,
-  },
-].map((v) => ({ ...v, aliases: [] }));
-
 async function promptForAffectedFlags(target: string) {
-  const telemetry = getTelemetry();
-  telemetry.featureUsed('affected-cli');
-
   const { positional, command, flags } = await selectAffectedFlags(target);
 
   if (flags !== undefined) {
@@ -235,60 +99,13 @@ async function promptForAffectedFlags(target: string) {
   }
 }
 
-/**
- *
- */
-async function selectAffectedFlags(target: string): Promise<{
-  command: string;
-  flags: string[] | undefined;
-  positional?: string;
-}> {
-  switch (target) {
-    case 'apps':
-    case 'libs': {
-      const command = `affected:${target}`;
-      return {
-        command,
-        flags: await selectFlags(command, AFFECTED_OPTIONS),
-      };
-    }
-    default: {
-      const positional = `--target=${target}`;
-      return {
-        command: 'affected',
-        positional,
-        flags: await selectFlags(`affected ${positional}`, AFFECTED_OPTIONS),
-      };
-    }
-  }
-}
-
 async function promptForRunMany() {
-  const telemetry = getTelemetry();
-  telemetry.featureUsed('affected-cli');
-
   const target = await promptForTarget();
   if (!target) {
     return;
   }
 
-  let options = RUN_MANY_OPTIONS;
-  const projects = await validProjectsForTarget(target);
-  if (projects && projects.length) {
-    options = [
-      {
-        name: 'projects',
-        type: OptionType.Array,
-        description: 'Projects to run',
-        aliases: [],
-        isRequired: false,
-        enum: projects,
-      },
-      ...RUN_MANY_OPTIONS,
-    ];
-  }
-
-  const flags = await selectFlags('run-many', options, { target });
+  const flags = await selectRunManyFlags(target);
 
   if (flags !== undefined) {
     const task = await NxTask.create({
@@ -300,8 +117,6 @@ async function promptForRunMany() {
 }
 
 async function promptForList() {
-  const telemetry = getTelemetry();
-  telemetry.featureUsed('affected-cli');
   const task = await NxTask.create({
     command: 'list',
     flags: [],
@@ -345,24 +160,6 @@ async function promptForMigrate() {
 
     tasks.executeTask(task);
   }
-}
-
-async function validProjectsForTarget(
-  target: string
-): Promise<string[] | undefined> {
-  const { validWorkspaceJson, workspace } = await getNxWorkspace();
-
-  if (!validWorkspaceJson || !workspace) {
-    return;
-  }
-
-  return Array.from(
-    new Set(
-      Object.entries(workspace.projects)
-        .filter(([, project]) => project.targets && project.targets[target])
-        .map(([project]) => project)
-    )
-  ).sort();
 }
 
 function buildQuickPickItems({
